Refactor FRIK code to folders by feature
