+++ src/FRIK.cpp
@@ -3,20 +3,20 @@
 #include <cpptrace/from_current.hpp>
 
 #include "Config.h"
-#include "ConfigurationMode.h"
 #include "Debug.h"
-#include "HandPose.h"
 #include "hook.h"
 #include "PapyrusApi.h"
 #include "PapyrusGateway.h"
 #include "patches.h"
-#include "Pipboy.h"
-#include "Skeleton.h"
-#include "SmoothMovementVR.h"
 #include "utils.h"
 #include "common/Logger.h"
+#include "config-mode/ConfigurationMode.h"
 #include "f4vr/F4VRUtils.h"
 #include "f4vr/VRControllersManager.h"
+#include "pipboy/Pipboy.h"
+#include "skeleton/HandPose.h"
+#include "skeleton/Skeleton.h"
+#include "smooth-movement/SmoothMovementVR.h"
 #include "ui/UIManager.h"
 #include "ui/UIModAdapter.h"
