+++ src/config-mode/ConfigurationMode.cpp
@@ -2,10 +2,10 @@
 
 #include "Config.h"
 #include "FRIK.h"
-#include "HandPose.h"
-#include "Skeleton.h"
 #include "f4vr/F4VRUtils.h"
 #include "f4vr/VRControllersManager.h"
+#include "skeleton/HandPose.h"
+#include "skeleton/Skeleton.h"
 
 using namespace common;
