+++ src/skeleton/BoneSpheresHandler.cpp
