+++ src/skeleton/CullGeometryHandler.cpp
