+++ src/weapon-position/WeaponPositionConfigMode.h
